+++ src/contexts/checkActiveEditor.ts
@@ -1,15 +1,15 @@
-import { commands, Disposable, window } from "vscode";
+import { Disposable, window } from "vscode";
 import { Status } from "../common/types";
 import { debounce } from "../decorators";
-import { Model } from "../model";
-import { IDisposable } from "../util";
+import { SourceControlManager } from "../source_control_manager";
+import { IDisposable, setVscodeContext } from "../util";
 
 export class CheckActiveEditor implements IDisposable {
   private disposables: Disposable[] = [];
 
-  constructor(private model: Model) {
+  constructor(private sourceControlManager: SourceControlManager) {
     // When repository update, like update
-    model.onDidChangeStatusRepository(
+    sourceControlManager.onDidChangeStatusRepository(
       this.checkHasChangesOnActiveEditor,
       this,
       this.disposables
@@ -24,8 +24,7 @@ export class CheckActiveEditor implements IDisposable {
 
   @debounce(100)
   private checkHasChangesOnActiveEditor() {
-    commands.executeCommand(
-      "setContext",
+    setVscodeContext(
       "svnActiveEditorHasChanges",
       this.hasChangesOnActiveEditor()
     );
@@ -37,7 +36,7 @@ export class CheckActiveEditor implements IDisposable {
     }
     const uri = window.activeTextEditor.document.uri;
 
-    const repository = this.model.getRepository(uri);
+    const repository = this.sourceControlManager.getRepository(uri);
     if (!repository) {
       return false;
     }