+++ src/contexts/openRepositoryCount.ts
@@ -1,25 +1,32 @@
-import { commands, Disposable } from "vscode";
+import { Disposable } from "vscode";
 import { debounce } from "../decorators";
-import { Model } from "../model";
-import { IDisposable } from "../util";
+import { SourceControlManager } from "../source_control_manager";
+import { IDisposable, setVscodeContext } from "../util";
 
 export class OpenRepositoryCount implements IDisposable {
   private disposables: Disposable[] = [];
 
-  constructor(private model: Model) {
+  constructor(private sourceControlManager: SourceControlManager) {
     // When repository Opened or closed
-    model.onDidOpenRepository(this.checkOpened, this, this.disposables);
-    model.onDidCloseRepository(this.checkOpened, this, this.disposables);
+    sourceControlManager.onDidOpenRepository(
+      this.checkOpened,
+      this,
+      this.disposables
+    );
+    sourceControlManager.onDidCloseRepository(
+      this.checkOpened,
+      this,
+      this.disposables
+    );
 
     this.checkOpened();
   }
 
   @debounce(100)
   private checkOpened() {
-    commands.executeCommand(
-      "setContext",
+    setVscodeContext(
       "svnOpenRepositoryCount",
-      `${this.model.repositories.length}`
+      this.sourceControlManager.repositories.length
     );
   }