+++ src/extension.ts
@@ -15,14 +15,14 @@ import { configuration } from "./helpers/configuration";
 import { ItemLogProvider } from "./historyView/itemLogProvider";
 import { RepoLogProvider } from "./historyView/repoLogProvider";
 import * as messages from "./messages";
-import { Model } from "./model";
+import { SourceControlManager } from "./source_control_manager";
 import { Svn } from "./svn";
 import { SvnContentProvider } from "./svnContentProvider";
 import { SvnFinder } from "./svnFinder";
 import SvnProvider from "./treeView/dataProviders/svnProvider";
-import {
-  toDisposable
-} from "./util";
+import { toDisposable } from "./util";
+import { BranchChangesProvider } from "./historyView/branchChangesProvider";
+import { IsSvn19orGreater } from "./contexts/isSvn19orGreater";
 
 async function init(
   _context: ExtensionContext,
@@ -34,27 +34,35 @@ async function init(
 
   const info = await svnFinder.findSvn(pathHint);
   const svn = new Svn({ svnPath: info.path, version: info.version });
-  const model = await new Model(svn, ConstructorPolicy.Async);
-  const contentProvider = new SvnContentProvider(model);
+  const sourceControlManager = await new SourceControlManager(
+    svn,
+    ConstructorPolicy.Async
+  );
+  const contentProvider = new SvnContentProvider(sourceControlManager);
 
-  registerCommands(model, disposables);
+  registerCommands(sourceControlManager, disposables);
 
-  disposables.push(model, contentProvider);
+  disposables.push(sourceControlManager, contentProvider);
 
-  const svnProvider = new SvnProvider(model);
+  const svnProvider = new SvnProvider(sourceControlManager);
 
   window.registerTreeDataProvider("svn", svnProvider);
 
-  const repoLogProvider = new RepoLogProvider(model);
+  const repoLogProvider = new RepoLogProvider(sourceControlManager);
   disposables.push(repoLogProvider);
   window.registerTreeDataProvider("repolog", repoLogProvider);
 
-  const itemLogProvider = new ItemLogProvider(model);
+  const itemLogProvider = new ItemLogProvider(sourceControlManager);
   disposables.push(itemLogProvider);
   window.registerTreeDataProvider("itemlog", itemLogProvider);
 
-  disposables.push(new CheckActiveEditor(model));
-  disposables.push(new OpenRepositoryCount(model));
+  const branchChangesProvider = new BranchChangesProvider(sourceControlManager);
+  disposables.push(branchChangesProvider);
+  window.registerTreeDataProvider("branchchanges", branchChangesProvider);
+
+  disposables.push(new CheckActiveEditor(sourceControlManager));
+  disposables.push(new OpenRepositoryCount(sourceControlManager));
+  disposables.push(new IsSvn19orGreater(info.version));
 
   outputChannel.appendLine(`Using svn "${info.version}" from "${info.path}"`);