+++ src/historyView/itemLogProvider.ts
@@ -12,7 +12,7 @@ import {
   window
 } from "vscode";
 import { ISvnLogEntry } from "../common/types";
-import { Model } from "../model";
+import { SourceControlManager } from "../source_control_manager";
 import { tempdir } from "../tempFiles";
 import { dispose, unwrap } from "../util";
 import {
@@ -29,7 +29,8 @@ import {
   LogTreeItemKind,
   openDiff,
   openFileRemote,
-  transform
+  transform,
+  getCommitDescription
 } from "./common";
 
 export class ItemLogProvider
@@ -43,7 +44,7 @@ export class ItemLogProvider
   private currentItem?: ICachedLog;
   private _dispose: Disposable[] = [];
 
-  constructor(private model: Model) {
+  constructor(private sourceControlManager: SourceControlManager) {
     window.onDidChangeActiveTextEditor(this.editorChanged, this);
     this._dispose.push(
       commands.registerCommand(
@@ -127,7 +128,7 @@ export class ItemLogProvider
         if (uri.path.startsWith(tempdir)) {
           return; // do not refresh if diff was called
         }
-        const repo = this.model.getRepository(uri);
+        const repo = this.sourceControlManager.getRepository(uri);
         if (repo !== null) {
           try {
             const info = await repo.getInfo(uri.fsPath);
@@ -156,6 +157,7 @@ export class ItemLogProvider
     if (element.kind === LogTreeItemKind.Commit) {
       const commit = element.data as ISvnLogEntry;
       ti = new TreeItem(getCommitLabel(commit), TreeItemCollapsibleState.None);
+      ti.description = getCommitDescription(commit);
       ti.iconPath = getCommitIcon(commit.author);
       ti.tooltip = getCommitToolTip(commit);
       ti.contextValue = "diffable";
@@ -182,6 +184,7 @@ export class ItemLogProvider
       const fname = path.basename(this.currentItem.svnTarget.fsPath);
       const ti = new TreeItem(fname, TreeItemCollapsibleState.Expanded);
       ti.tooltip = path.dirname(this.currentItem.svnTarget.fsPath);
+      ti.description = path.dirname(this.currentItem.svnTarget.fsPath);
       ti.iconPath = getIconObject("icon-history");
       const item = {
         kind: LogTreeItemKind.TItem,