+++ src/historyView/repoLogProvider.ts
@@ -12,12 +12,12 @@ import {
   workspace
 } from "vscode";
 import {
-  IModelChangeEvent,
+  RepositoryChangeEvent,
   ISvnLogEntry,
   ISvnLogEntryPath
 } from "../common/types";
 import { exists } from "../fs";
-import { Model } from "../model";
+import { SourceControlManager } from "../source_control_manager";
 import { IRemoteRepository } from "../remoteRepository";
 import { Repository } from "../repository";
 import { dispose, unwrap } from "../util";
@@ -37,7 +37,8 @@ import {
   openDiff,
   openFileRemote,
   SvnPath,
-  transform
+  transform,
+  getCommitDescription
 } from "./common";
 
 function getActionIcon(action: string) {
@@ -81,7 +82,7 @@ export class RepoLogProvider
     return this.getCached(item.parent);
   }
 
-  constructor(private model: Model) {
+  constructor(private sourceControlManager: SourceControlManager) {
     this.refresh();
     this._dispose.push(
       commands.registerCommand(
@@ -119,10 +120,12 @@ export class RepoLogProvider
     this._dispose.push(
       commands.registerCommand("svn.repolog.refresh", this.refresh, this)
     );
-    this.model.onDidChangeRepository(async (_e: IModelChangeEvent) => {
-      return this.refresh();
-      // TODO refresh only required repo, need to pass element === getChildren()
-    });
+    this.sourceControlManager.onDidChangeRepository(
+      async (_e: RepositoryChangeEvent) => {
+        return this.refresh();
+        // TODO refresh only required repo, need to pass element === getChildren()
+      }
+    );
   }
 
   public dispose() {
@@ -151,12 +154,12 @@ export class RepoLogProvider
       window.showWarningMessage("This path is already added");
       return;
     }
-    const repo = this.model.getRepository(repoLike);
+    const repo = this.sourceControlManager.getRepository(repoLike);
     if (repo === null) {
       try {
         let uri: Uri;
         if (repoLike.startsWith("^")) {
-          const wsrepo = this.model.getRepository(
+          const wsrepo = this.sourceControlManager.getRepository(
             unwrap(workspace.workspaceFolders)[0].uri
           );
           if (!wsrepo) {
@@ -170,7 +173,9 @@ export class RepoLogProvider
         if (rev !== "HEAD" && isNaN(parseInt(rev, 10))) {
           throw new Error("erroneous revision");
         }
-        const remRepo = await this.model.getRemoteRepository(uri);
+        const remRepo = await this.sourceControlManager.getRemoteRepository(
+          uri
+        );
         item.repo = remRepo;
         item.svnTarget = uri;
       } catch (e) {
@@ -313,7 +318,7 @@ export class RepoLogProvider
           this.logCache.delete(k);
         }
       }
-      for (const repo of this.model.repositories) {
+      for (const repo of this.sourceControlManager.repositories) {
         const remoteRoot = repo.branchRoot;
         const repoUrl = remoteRoot.toString(true);
         let persisted: ICachedLog["persisted"] = {
@@ -373,6 +378,7 @@ export class RepoLogProvider
         getCommitLabel(commit),
         TreeItemCollapsibleState.Collapsed
       );
+      ti.description = getCommitDescription(commit);
       ti.tooltip = getCommitToolTip(commit);
       ti.iconPath = getCommitIcon(commit.author);
       ti.contextValue = "commit";
@@ -381,6 +387,7 @@ export class RepoLogProvider
       const pathElem = element.data as ISvnLogEntryPath;
       const basename = path.basename(pathElem._);
       ti = new TreeItem(basename, TreeItemCollapsibleState.None);
+      ti.description = path.dirname(pathElem._);
       const cached = this.getCached(element);
       const nm = cached.repo.getPathNormalizer();
       ti.tooltip = nm.parse(pathElem._).relativeFromBranch;