+++ src/test/testUtil.ts
@@ -185,7 +185,10 @@ export function activeExtension() {
     }
 
     if (!extension.isActive) {
-      extension.activate().then(() => resolve(), () => reject());
+      extension.activate().then(
+        () => resolve(),
+        () => reject()
+      );
     } else {
       resolve();
     }