+++ CHANGELOG.md
@@ -1,3 +1,73 @@
+# [2.4.0](https://github.com/JohnstonCode/svn-scm/compare/v2.3.0...v2.4.0) (2020-01-07)
+
+
+### Features
+
+* Add remove unversioned command ([#769](https://github.com/JohnstonCode/svn-scm/issues/769)) ([177bb0b](https://github.com/JohnstonCode/svn-scm/commit/177bb0baf88cf01b618503ce4c8c2a1569ec5ea4))
+
+# [2.3.0](https://github.com/JohnstonCode/svn-scm/compare/v2.2.1...v2.3.0) (2020-01-01)
+
+
+### Features
+
+* New svn icon ([#758](https://github.com/JohnstonCode/svn-scm/issues/758)) ([c1c20b4](https://github.com/JohnstonCode/svn-scm/commit/c1c20b4))
+
+## [2.2.1](https://github.com/JohnstonCode/svn-scm/compare/v2.2.0...v2.2.1) (2019-12-27)
+
+
+### Bug Fixes
+
+* Temp files are encoded using default encoding setting ([#765](https://github.com/JohnstonCode/svn-scm/issues/765)) ([6c2748a](https://github.com/JohnstonCode/svn-scm/commit/6c2748a))
+
+# [2.2.0](https://github.com/JohnstonCode/svn-scm/compare/v2.1.3...v2.2.0) (2019-12-27)
+
+
+### Features
+
+* Added pretty descriptions to treeview ([#759](https://github.com/JohnstonCode/svn-scm/issues/759)) ([73de872](https://github.com/JohnstonCode/svn-scm/commit/73de872))
+
+## [2.1.3](https://github.com/JohnstonCode/svn-scm/compare/v2.1.2...v2.1.3) (2019-11-16)
+
+
+### Bug Fixes
+
+* Added check for jschardet constants ([#745](https://github.com/JohnstonCode/svn-scm/issues/745)) ([7594012](https://github.com/JohnstonCode/svn-scm/commit/7594012))
+
+## [2.1.2](https://github.com/JohnstonCode/svn-scm/compare/v2.1.1...v2.1.2) (2019-11-15)
+
+
+### Bug Fixes
+
+* Debounce repo watcher ([#742](https://github.com/JohnstonCode/svn-scm/issues/742)) ([a34b79b](https://github.com/JohnstonCode/svn-scm/commit/a34b79b))
+
+## [2.1.1](https://github.com/JohnstonCode/svn-scm/compare/v2.1.0...v2.1.1) (2019-11-14)
+
+
+### Bug Fixes
+
+* Fix cwd when running svn commands for remoteRepository ([#738](https://github.com/JohnstonCode/svn-scm/issues/738)) ([2c0dd2d](https://github.com/JohnstonCode/svn-scm/commit/2c0dd2d))
+
+# [2.1.0](https://github.com/JohnstonCode/svn-scm/compare/v2.0.2...v2.1.0) (2019-11-13)
+
+
+### Features
+
+* Added Branches tree view ([#729](https://github.com/JohnstonCode/svn-scm/issues/729)) ([cecd185](https://github.com/JohnstonCode/svn-scm/commit/cecd185))
+
+## [2.0.2](https://github.com/JohnstonCode/svn-scm/compare/v2.0.1...v2.0.2) (2019-11-11)
+
+
+### Bug Fixes
+
+* Fixed CSP policy for commit page ([#733](https://github.com/JohnstonCode/svn-scm/issues/733)) ([6996ff0](https://github.com/JohnstonCode/svn-scm/commit/6996ff0))
+
+## [2.0.1](https://github.com/JohnstonCode/svn-scm/compare/v2.0.0...v2.0.1) (2019-11-10)
+
+
+### Bug Fixes
+
+* Stop retrieving info of repo root ([#728](https://github.com/JohnstonCode/svn-scm/issues/728)) ([12e584b](https://github.com/JohnstonCode/svn-scm/commit/12e584b))
+
 # [2.0.0](https://github.com/JohnstonCode/svn-scm/compare/v1.54.11...v2.0.0) (2019-11-08)